Added some configuration options to the dpgit module where the path to git can be specified.
Added _dpgit_shell_command that can be used to construct a git shell command with the correct path, work dir and git dir. The returned string can then be executed with exec or shell_exec. Rewrote all methods that used exec to interact with git to use this helper function.

Changes in DProjectRelease. Added accessor functions for the rid and project node attributes. Implemented the save function. The static load function now loads the new packager and created columns too. Some changes and bugfixes in the version handling functions.

Numerous changes in the release creation process. The release type terms are now set as metadata on the release. The release is now persisted in the database when the create release form is submitted, and only the rid is sent to identify the release for the confirmation page.

The changed the default value for minor_version@dproject_release to NULL.